+++ server/app.js
@@ -4,15 +4,12 @@ const { join } = require('path');
 const cookieParser = require('cookie-parser');
 const logger = require('morgan');
 const passport = require('passport');
-
-const indexRouter = require('./routes/index');
-const pingRouter = require('./routes/ping');
-const userRouter = require('./routes/user');
+const routes = require('./routes');
+const passportConfig = require('./config/passport-config');
 
 // DB connection
 const connectDB = require('./db/connection');
 
-const passportConfig = require('./config/passport-config');
 const { json, urlencoded } = express;
 
 const app = express();
@@ -30,9 +27,7 @@ app.use(urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(join(__dirname, 'public')));
 
-app.use('/', indexRouter);
-app.use('/ping', pingRouter);
-app.use('/user', userRouter);
+app.use(routes);
 
 // catch 404 and forward to error handler
 app.use((req, res, next) => {
